// private/zconcert-admin/types/ElementObject.ts
export type SpanElements = {
    sandyA: HTMLSpanElement;
    sandyB: HTMLSpanElement;
    sandyC: HTMLSpanElement;
    sandyD: HTMLSpanElement;
};

export type InputElements = {
    sandyA: HTMLInputElement;
    sandyB: HTMLInputElement;
    sandyC: HTMLInputElement;
    sandyD: HTMLInputElement;
};